Stop using the DefaultServeMux
This change uses a custom ServeMux instead of Go's default. The reason
for this is the following, taken from Let's Go by Alex Edwards:

Because DefaultServeMux is a global variable, any package can access it
and register a route — including any third-party packages that your
application imports. If one of those third- party packages is
compromised, they could use DefaultServeMux to expose a malicious
handler to the web.

So, for the sake of security, it’s generally a good idea to avoid
DefaultServeMux and the corresponding helper functions. Use your own
locally-scoped servemux instead, like we have been doing in this project
so far.

Grab a copy of the book at https://lets-go.alexedwards.net/, if you're
keen.